+++ src/fields/nonnative/field_var.rs
@@ -1,4 +1,4 @@
-use super::{params::OptimizationType, AllocatedNonNativeFieldVar, NonNativeFieldMulResultVar};
+use super::{params::OptimizationType, AllocatedEmulatedFpVar, MulResultVar};
 use crate::{
     boolean::Boolean,
     fields::{fp::FpVar, FieldVar},
@@ -17,15 +17,15 @@ use ark_std::{
 /// constraint field (`BaseField`).
 #[derive(Clone, Debug)]
 #[must_use]
-pub enum NonNativeFieldVar<TargetField: PrimeField, BaseField: PrimeField> {
+pub enum EmulatedFpVar<TargetField: PrimeField, BaseField: PrimeField> {
     /// Constant
     Constant(TargetField),
     /// Allocated gadget
-    Var(AllocatedNonNativeFieldVar<TargetField, BaseField>),
+    Var(AllocatedEmulatedFpVar<TargetField, BaseField>),
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> PartialEq
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     fn eq(&self, other: &Self) -> bool {
         self.value()
@@ -35,12 +35,12 @@ impl<TargetField: PrimeField, BaseField: PrimeField> PartialEq
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> Eq
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> Hash
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     fn hash<H: Hasher>(&self, state: &mut H) {
         self.value().unwrap_or_default().hash(state);
@@ -48,7 +48,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Hash
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> R1CSVar<BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     type Value = TargetField;
 
@@ -68,7 +68,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> R1CSVar<BaseField>
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> From<Boolean<BaseField>>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     fn from(other: Boolean<BaseField>) -> Self {
         if let Boolean::Constant(b) = other {
@@ -83,27 +83,27 @@ impl<TargetField: PrimeField, BaseField: PrimeField> From<Boolean<BaseField>>
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField>
-    From<AllocatedNonNativeFieldVar<TargetField, BaseField>>
-    for NonNativeFieldVar<TargetField, BaseField>
+    From<AllocatedEmulatedFpVar<TargetField, BaseField>>
+    for EmulatedFpVar<TargetField, BaseField>
 {
-    fn from(other: AllocatedNonNativeFieldVar<TargetField, BaseField>) -> Self {
+    fn from(other: AllocatedEmulatedFpVar<TargetField, BaseField>) -> Self {
         Self::Var(other)
     }
 }
 
 impl<'a, TargetField: PrimeField, BaseField: PrimeField> FieldOpsBounds<'a, TargetField, Self>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
 }
 
 impl<'a, TargetField: PrimeField, BaseField: PrimeField>
-    FieldOpsBounds<'a, TargetField, NonNativeFieldVar<TargetField, BaseField>>
-    for &'a NonNativeFieldVar<TargetField, BaseField>
+    FieldOpsBounds<'a, TargetField, EmulatedFpVar<TargetField, BaseField>>
+    for &'a EmulatedFpVar<TargetField, BaseField>
 {
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> FieldVar<TargetField, BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     fn zero() -> Self {
         Self::Constant(TargetField::zero())
@@ -147,33 +147,33 @@ impl<TargetField: PrimeField, BaseField: PrimeField> FieldVar<TargetField, BaseF
 }
 
 impl_bounded_ops!(
-    NonNativeFieldVar<TargetField, BaseField>,
+    EmulatedFpVar<TargetField, BaseField>,
     TargetField,
     Add,
     add,
     AddAssign,
     add_assign,
-    |this: &'a NonNativeFieldVar<TargetField, BaseField>, other: &'a NonNativeFieldVar<TargetField, BaseField>| {
-        use NonNativeFieldVar::*;
+    |this: &'a EmulatedFpVar<TargetField, BaseField>, other: &'a EmulatedFpVar<TargetField, BaseField>| {
+        use EmulatedFpVar::*;
         match (this, other) {
             (Constant(c1), Constant(c2)) => Constant(*c1 + c2),
             (Constant(c), Var(v)) | (Var(v), Constant(c)) => Var(v.add_constant(c).unwrap()),
             (Var(v1), Var(v2)) => Var(v1.add(v2).unwrap()),
         }
     },
-    |this: &'a NonNativeFieldVar<TargetField, BaseField>, other: TargetField| { this + &NonNativeFieldVar::Constant(other) },
+    |this: &'a EmulatedFpVar<TargetField, BaseField>, other: TargetField| { this + &EmulatedFpVar::Constant(other) },
     (TargetField: PrimeField, BaseField: PrimeField),
 );
 
 impl_bounded_ops!(
-    NonNativeFieldVar<TargetField, BaseField>,
+    EmulatedFpVar<TargetField, BaseField>,
     TargetField,
     Sub,
     sub,
     SubAssign,
     sub_assign,
-    |this: &'a NonNativeFieldVar<TargetField, BaseField>, other: &'a NonNativeFieldVar<TargetField, BaseField>| {
-        use NonNativeFieldVar::*;
+    |this: &'a EmulatedFpVar<TargetField, BaseField>, other: &'a EmulatedFpVar<TargetField, BaseField>| {
+        use EmulatedFpVar::*;
         match (this, other) {
             (Constant(c1), Constant(c2)) => Constant(*c1 - c2),
             (Var(v), Constant(c)) => Var(v.sub_constant(c).unwrap()),
@@ -181,32 +181,32 @@ impl_bounded_ops!(
             (Var(v1), Var(v2)) => Var(v1.sub(v2).unwrap()),
         }
     },
-    |this: &'a NonNativeFieldVar<TargetField, BaseField>, other: TargetField| {
-        this - &NonNativeFieldVar::Constant(other)
+    |this: &'a EmulatedFpVar<TargetField, BaseField>, other: TargetField| {
+        this - &EmulatedFpVar::Constant(other)
     },
     (TargetField: PrimeField, BaseField: PrimeField),
 );
 
 impl_bounded_ops!(
-    NonNativeFieldVar<TargetField, BaseField>,
+    EmulatedFpVar<TargetField, BaseField>,
     TargetField,
     Mul,
     mul,
     MulAssign,
     mul_assign,
-    |this: &'a NonNativeFieldVar<TargetField, BaseField>, other: &'a NonNativeFieldVar<TargetField, BaseField>| {
-        use NonNativeFieldVar::*;
+    |this: &'a EmulatedFpVar<TargetField, BaseField>, other: &'a EmulatedFpVar<TargetField, BaseField>| {
+        use EmulatedFpVar::*;
         match (this, other) {
             (Constant(c1), Constant(c2)) => Constant(*c1 * c2),
             (Constant(c), Var(v)) | (Var(v), Constant(c)) => Var(v.mul_constant(c).unwrap()),
             (Var(v1), Var(v2)) => Var(v1.mul(v2).unwrap()),
         }
     },
-    |this: &'a NonNativeFieldVar<TargetField, BaseField>, other: TargetField| {
+    |this: &'a EmulatedFpVar<TargetField, BaseField>, other: TargetField| {
         if other.is_zero() {
-            NonNativeFieldVar::zero()
+            EmulatedFpVar::zero()
         } else {
-            this * &NonNativeFieldVar::Constant(other)
+            this * &EmulatedFpVar::Constant(other)
         }
     },
     (TargetField: PrimeField, BaseField: PrimeField),
@@ -216,7 +216,7 @@ impl_bounded_ops!(
 /// *************************************************************************
 
 impl<TargetField: PrimeField, BaseField: PrimeField> EqGadget<BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     #[tracing::instrument(target = "r1cs")]
     fn is_eq(&self, other: &Self) -> R1CSResult<Boolean<BaseField>> {
@@ -250,7 +250,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> EqGadget<BaseField>
             },
             (Self::Constant(c), Self::Var(v)) | (Self::Var(v), Self::Constant(c)) => {
                 let cs = v.cs();
-                let c = AllocatedNonNativeFieldVar::new_constant(cs, c)?;
+                let c = AllocatedEmulatedFpVar::new_constant(cs, c)?;
                 c.conditional_enforce_equal(v, should_enforce)
             },
             (Self::Var(v1), Self::Var(v2)) => v1.conditional_enforce_equal(v2, should_enforce),
@@ -272,7 +272,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> EqGadget<BaseField>
             },
             (Self::Constant(c), Self::Var(v)) | (Self::Var(v), Self::Constant(c)) => {
                 let cs = v.cs();
-                let c = AllocatedNonNativeFieldVar::new_constant(cs, c)?;
+                let c = AllocatedEmulatedFpVar::new_constant(cs, c)?;
                 c.conditional_enforce_not_equal(v, should_enforce)
             },
             (Self::Var(v1), Self::Var(v2)) => v1.conditional_enforce_not_equal(v2, should_enforce),
@@ -281,7 +281,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> EqGadget<BaseField>
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> ToBitsGadget<BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     #[tracing::instrument(target = "r1cs")]
     fn to_bits_le(&self) -> R1CSResult<Vec<Boolean<BaseField>>> {
@@ -305,7 +305,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> ToBitsGadget<BaseField>
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> ToBytesGadget<BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     /// Outputs the unique byte decomposition of `self` in *little-endian*
     /// form.
@@ -332,7 +332,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> ToBytesGadget<BaseField>
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> CondSelectGadget<BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     #[tracing::instrument(target = "r1cs")]
     fn conditionally_select(
@@ -346,11 +346,11 @@ impl<TargetField: PrimeField, BaseField: PrimeField> CondSelectGadget<BaseField>
             _ => {
                 let cs = cond.cs();
                 let true_value = match true_value {
-                    Self::Constant(f) => AllocatedNonNativeFieldVar::new_constant(cs.clone(), f)?,
+                    Self::Constant(f) => AllocatedEmulatedFpVar::new_constant(cs.clone(), f)?,
                     Self::Var(v) => v.clone(),
                 };
                 let false_value = match false_value {
-                    Self::Constant(f) => AllocatedNonNativeFieldVar::new_constant(cs, f)?,
+                    Self::Constant(f) => AllocatedEmulatedFpVar::new_constant(cs, f)?,
                     Self::Var(v) => v.clone(),
                 };
                 cond.select(&true_value, &false_value).map(Self::Var)
@@ -362,7 +362,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> CondSelectGadget<BaseField>
 /// Uses two bits to perform a lookup into a table
 /// `b` is little-endian: `b[0]` is LSB.
 impl<TargetField: PrimeField, BaseField: PrimeField> TwoBitLookupGadget<BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     type TableConstant = TargetField;
 
@@ -378,13 +378,13 @@ impl<TargetField: PrimeField, BaseField: PrimeField> TwoBitLookupGadget<BaseFiel
             let index = lsb + (msb << 1);
             Ok(Self::Constant(c[index]))
         } else {
-            AllocatedNonNativeFieldVar::two_bit_lookup(b, c).map(Self::Var)
+            AllocatedEmulatedFpVar::two_bit_lookup(b, c).map(Self::Var)
         }
     }
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> ThreeBitCondNegLookupGadget<BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     type TableConstant = TargetField;
 
@@ -413,13 +413,13 @@ impl<TargetField: PrimeField, BaseField: PrimeField> ThreeBitCondNegLookupGadget
             };
             Ok(Self::Constant(y))
         } else {
-            AllocatedNonNativeFieldVar::three_bit_cond_neg_lookup(b, b0b1, c).map(Self::Var)
+            AllocatedEmulatedFpVar::three_bit_cond_neg_lookup(b, b0b1, c).map(Self::Var)
         }
     }
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> AllocVar<TargetField, BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     fn new_variable<T: Borrow<TargetField>>(
         cs: impl Into<Namespace<BaseField>>,
@@ -432,13 +432,13 @@ impl<TargetField: PrimeField, BaseField: PrimeField> AllocVar<TargetField, BaseF
         if cs == ConstraintSystemRef::None || mode == AllocationMode::Constant {
             Ok(Self::Constant(*f()?.borrow()))
         } else {
-            AllocatedNonNativeFieldVar::new_variable(cs, f, mode).map(Self::Var)
+            AllocatedEmulatedFpVar::new_variable(cs, f, mode).map(Self::Var)
         }
     }
 }
 
 impl<TargetField: PrimeField, BaseField: PrimeField> ToConstraintFieldGadget<BaseField>
-    for NonNativeFieldVar<TargetField, BaseField>
+    for EmulatedFpVar<TargetField, BaseField>
 {
     #[tracing::instrument(target = "r1cs")]
     fn to_constraint_field(&self) -> R1CSResult<Vec<FpVar<BaseField>>> {
@@ -447,7 +447,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> ToConstraintFieldGadget<Bas
         // By default, the constant is converted in the weight-optimized type, because
         // it results in fewer elements.
         match self {
-            Self::Constant(c) => Ok(AllocatedNonNativeFieldVar::get_limbs_representations(
+            Self::Constant(c) => Ok(AllocatedEmulatedFpVar::get_limbs_representations(
                 c,
                 OptimizationType::Weight,
             )?
@@ -459,23 +459,23 @@ impl<TargetField: PrimeField, BaseField: PrimeField> ToConstraintFieldGadget<Bas
     }
 }
 
-impl<TargetField: PrimeField, BaseField: PrimeField> NonNativeFieldVar<TargetField, BaseField> {
-    /// The `mul_without_reduce` for `NonNativeFieldVar`
+impl<TargetField: PrimeField, BaseField: PrimeField> EmulatedFpVar<TargetField, BaseField> {
+    /// The `mul_without_reduce` for `EmulatedFpVar`
     #[tracing::instrument(target = "r1cs")]
     pub fn mul_without_reduce(
         &self,
         other: &Self,
-    ) -> R1CSResult<NonNativeFieldMulResultVar<TargetField, BaseField>> {
+    ) -> R1CSResult<MulResultVar<TargetField, BaseField>> {
         match self {
             Self::Constant(c) => match other {
-                Self::Constant(other_c) => Ok(NonNativeFieldMulResultVar::Constant(*c * other_c)),
+                Self::Constant(other_c) => Ok(MulResultVar::Constant(*c * other_c)),
                 Self::Var(other_v) => {
                     let self_v =
-                        AllocatedNonNativeFieldVar::<TargetField, BaseField>::new_constant(
+                        AllocatedEmulatedFpVar::<TargetField, BaseField>::new_constant(
                             self.cs(),
                             c,
                         )?;
-                    Ok(NonNativeFieldMulResultVar::Var(
+                    Ok(MulResultVar::Var(
                         other_v.mul_without_reduce(&self_v)?,
                     ))
                 },
@@ -483,14 +483,14 @@ impl<TargetField: PrimeField, BaseField: PrimeField> NonNativeFieldVar<TargetFie
             Self::Var(v) => {
                 let other_v = match other {
                     Self::Constant(other_c) => {
-                        AllocatedNonNativeFieldVar::<TargetField, BaseField>::new_constant(
+                        AllocatedEmulatedFpVar::<TargetField, BaseField>::new_constant(
                             self.cs(),
                             other_c,
                         )?
                     },
                     Self::Var(other_v) => other_v.clone(),
                 };
-                Ok(NonNativeFieldMulResultVar::Var(
+                Ok(MulResultVar::Var(
                     v.mul_without_reduce(&other_v)?,
                 ))
             },